requirements.txt
PySide6==6.8.1
pyqtgraph==0.13.7
pyserial==3.5
numpy==2.2.4
scipy==1.15.2
pytest==8.3.5
